data scrapping working auto completed
